+++ packages/compiler/src/component.ts
